fixed notice in autoupgrade install + check if deleteDirectory function exists

+++ modules/autoupgrade/autoupgrade.php
@@ -20,7 +20,7 @@
 *
 *  @author PrestaShop SA <contact@prestashop.com>
 *  @copyright  2007-2011 PrestaShop SA
-*  @version  Release: $Revision: 10460 $
+*  @version  Release: $Revision: 13523 $
 *  @license    http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
 *  International Registered Trademark & Property of PrestaShop SA
 */
@@ -63,7 +63,7 @@ class Autoupgrade extends Module
 		{
 			$tab = new Tab($idTab);
 			// this deletion will not cancel the installation
-			$resDelete &= $tab->delete();
+			$resDelete = $tab->delete();
 			if (!$resDelete)
 				$this->_errors[] = sprintf($this->l('Unable to delete outdated AdminUpgrade tab %s'), $idTab);
 		}
@@ -143,6 +143,7 @@ class Autoupgrade extends Module
 
 		return true;
 	}
+
 	public function uninstall()
 	{
 		$id_tab = Configuration::get('PS_AUTOUPDATE_MODULE_IDTAB');
@@ -165,14 +166,14 @@ class Autoupgrade extends Module
 			$res &= unlink(_PS_ADMIN_DIR_.DIRECTORY_SEPARATOR.'tabs'.'AdminUpgrade.php');
 		}
 		
-		// there is no return value in Tools::deleteDirectory
-		Tools::deleteDirectory(_PS_ADMIN_DIR_.DIRECTORY_SEPARATOR.'autoupgrade', false);
+		// if the function does not exists, ignore it
+		// (there is no return value in Tools::deleteDirectory)
+		if (method_exists('Tools', 'deleteDirectory'))
+			Tools::deleteDirectory(_PS_ADMIN_DIR_.DIRECTORY_SEPARATOR.'autoupgrade', false);
 
 		if (!$res OR !parent::uninstall())
 			return false;
 
 		return true;
 	}
-
-
 }
